하이버네이트 오류 수정

+++ erp_mes/src/main/java/com/erp_mes/mes/quality/dto/InspectionItemDTO.java
@@ -1,5 +1,7 @@
 package com.erp_mes.mes.quality.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import lombok.Setter;
 
@@ -9,6 +11,6 @@ public class InspectionItemDTO {
     private Long itemId;
     private String productId; // 제품번호
     private Long inspectionFMId; // INSPECTION_FM_ID 검사기준값
-    private Double toleranceValue; // 허용 공차
+    private BigDecimal toleranceValue; // 허용 공차
     private String unit;
 }
